Ticket 54: was andere Leser:innen sagen
Die Onleihe nennt auf jeder Detailseite einen Durchschnitt und die Zahl
der Stimmen — auf derselben Seite, die ohnehin fuer die Verfuegbarkeit
geholt wird. Gemessen: 22 bis 1641 Stimmen je Titel.

**Keine eigene Tabelle.** `rating` ist seit ADR 19 nach
`(subject, origin)` geschluesselt, und genau darauf kommt eine weitere
Quelle dazu: die Bibliotheksleserschaft ist die vierte Herkunft neben
Modell, Gespraech und Leserin. Mein erster Entwurf baute dieselbe Achse
ein zweites Mal als Tabelle `reception`.

`stars` traegt jetzt Nachkommastellen, weil fremde Stimmen sie
mitbringen; eigene Urteile bleiben ganzzahlig, das erzwingt das Schema.
Ein Jinja-Filter schreibt "4" statt "4.0".

`profile_version` ist 0: eine fremde Note ist kein Urteil gegen das
Leseprofil, veraltet also nicht mit einer neuen Fassung, und `ebw rate`
schlaegt ohnehin `(subject, BY_MODEL)` nach. Auf der Buchseite steht sie
in einem eigenen Abschnitt — eine Zahl statt Sternsymbolen, mit der
Stimmenzahl daneben und ohne "veraltet"-Abzeichen.

Ohne Stimmenzahl wird nichts geschrieben: gemessen an Google Books
ruhen fuenf von sieben Bewertungen unseres Korpus auf einer einzigen
Stimme (docs/research/reader-ratings-sources.md, neu).

Dazu zwei Korrekturen an eigenen Fehlern:

Migration 16 wurde von mir **ersetzt** statt angehaengt — der Fehler,
den ADR 16 ausdruecklich verbietet. Eine Datei, die sie schon gelaufen
war, ueberspringt den Nachfolger fuer immer; genau das hat
`rating.votes` bei `user_version = 17` fehlen lassen. 16 bleibt jetzt
als dokumentierte Sackgasse stehen, der Aufraeumer ist angehaengt.

Die frueher gemessene Cache-Vorgabe von Google Books (max-age=86400)
galt einem Titelbild. Die JSON-Antwort traegt gar keinen Cache-Header —
das Ausschluss-Argument liess sich nicht uebertragen, und die
Rechtslage ist ohnehin nicht meine Entscheidung.

+++ src/ebook_watchlist/rating.py
@@ -73,7 +73,9 @@ class RatingUnavailable(Exception):
 
 @dataclass(frozen=True, slots=True)
 class Rating:
-    stars: int
+    #: Ganzzahlig bei eigenen Urteilen; fremde Stimmen bringen
+    #: Nachkommastellen mit (Ticket 54).
+    stars: float
     reason: str
     #: ``belegt`` | ``teils`` | ``vermutet`` — das Schema verlangt sie, weil ein
     #: Urteil über einen 219 Zeichen langen Anriss etwas anderes ist als eines
